requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
az-iranian-bank-gateways==2.0.5
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.19.0
boto3==1.34.151
botocore==1.34.151
bs4==0.0.2
celery==5.4.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
datetime-truncate==1.1.1
Django==5.0.6
django-admin-charts==1.4.0
django-allauth==64.0.0
django-bower==5.2.0
django-celery-beat==2.6.0
django-ckeditor==6.6.1
django-ckeditor-5==0.2.13
django-htmx==1.19.0
django-jalali==7.0.0
django-js-asset==2.2.0
django-light==0.1.0.post3
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-render-partial==0.4
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-swagger==0.2.2
drf-spectacular==0.27.2
fire==0.4.0
ghasedak_sms==1.0.0
ghasedakpack==0.1.13
gunicorn==23.0.0
halo==0.0.31
html5lib==1.1
hyperlink==21.0.0
incremental==24.7.2
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
log-symbols==0.0.14
lxml==5.2.2
mailchimp-marketing==3.0.80
MarkupSafe==2.1.5
msgpack==1.0.8
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==24.1
pathspec==0.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.10.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
readchar==2.0.1
redis==5.0.8
referencing==0.35.1
regex==2023.6.3
reportlab==4.2.2
requests-file==2.1.0
rest-framework-simplejwt==0.0.2
rpds-py==0.19.1
s3transfer==0.10.2
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
spinners==0.0.24
sqlparse==0.5.1
svglib==1.5.1
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.3.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
uritools==4.0.3
vine==5.1.0
webencodings==0.5.1
websockets==10.0
xhtml2pdf==0.2.16
zeep==4.2.1
zope.interface==6.4.post2
django-template-maths==0.2.0
django-cors-headers==4.4.0